Start of GitHub action
This is a start of turning the escript into a reusable action. To make
the Dockerbuild a bit faster this also adds a rebar.config.script that
disables the plugins, as they otherwise would be downloaded and compiled
even before `rebar get-deps`.

To make it a bit easier to test locally it includes a small shell
script, priv/build-and-run.sh. It passes all extra arguments to the
escript, for example `./priv/build-and-run.sh --help`.